+++ docs/changelog.md
@@ -1,3 +1,7 @@
 # CHANGELOG
 
 ## develop branch
+
+### New
+
+* `build-contracts` - convert PHPUnit testdox.txt into Markdown